FAC-770: refactor - Rename `id` to `kitCustomId` in `UseCase`
